fmt: move pages background in parent directory
